% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{get_rnaseq_diff_expr_data}
\alias{get_rnaseq_diff_expr_data}
\title{Get rnaseq differential expression data from Synapse}
\usage{
get_rnaseq_diff_expr_data(id, models_to_keep)
}
\description{
Get rnaseq differential expression data from Synapse
}
