% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{process_brain_expression_data}
\alias{process_brain_expression_data}
\title{Process brain expression data}
\usage{
process_brain_expression_data(data, gene_info, fdr_random_threshold)
}
\description{
Process brain expression data
}
