% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{process_median_expression_data}
\alias{process_median_expression_data}
\title{Process median expression data}
\usage{
process_median_expression_data(data, gene_info)
}
\description{
Process median expression data
}
