% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{process_rnaseq_diff_expr_data}
\alias{process_rnaseq_diff_expr_data}
\title{Process rnaseq diff expr data}
\usage{
process_rnaseq_diff_expr_data(data, gene_info, adj_p_value_threshold,
  target_list)
}
\description{
Process rnaseq diff expr data
}
