% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{process_target_list}
\alias{process_target_list}
\title{Process the target list by getting specific columns,
splitting Synapse IDs for data used, and grouping
and nesting by Ensembl gene id into nominatedtarget column.}
\usage{
process_target_list(data)
}
\description{
Process the target list by getting specific columns,
splitting Synapse IDs for data used, and grouping
and nesting by Ensembl gene id into nominatedtarget column.
}
