% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{process_team}
\alias{process_team}
\title{Process team info by joining with team member info.}
\usage{
process_team(team_info, team_member_info)
}
\description{
Process team info by joining with team member info.
}
